lat/lng to float for accuracy fix

+++ db/migrate/20190222211149_add_lat_lng_to_users.rb
@@ -1,6 +1,6 @@
 class AddLatLngToUsers < ActiveRecord::Migration[5.2]
   def change
-    add_column :users, :lat, :integer
-    add_column :users, :lng, :integer
+    add_column :users, :lat, :float
+    add_column :users, :lng, :float
   end
 end
