+++ db/migrate/20190223235233_create_gyms.rb
@@ -1,8 +1,8 @@
 class CreateGyms < ActiveRecord::Migration[5.2]
   def change
     create_table :gyms do |t|
-      t.integer :lat
-      t.integer :lng
+      t.float :lat
+      t.float :lng
       t.string :name, null: false
     end
   end
